fix: resolve material plugin error when remove block module
